refactor(fetch): extract duplicate keyword filters into shared traits
The four Fetch ability handlers (FetchRss, FetchWordPressApi,
FetchWordPressMedia, QueryWordPressPosts) each carried byte-identical copies of
applyKeywordSearch and applyKeywordExclusion — a comma-separated keyword
matcher and its negation. Hoist both into shared traits under
DataMachine\Abilities\Fetch\Traits and compose them into each handler.

- New HasApplyKeywordSearch + HasApplyKeywordExclusion traits.
- Each handler drops the two duplicated methods and adds the two trait uses.
- Net -104 lines (188 removed, 84 added) across 6 files; call sites
  ($this->applyKeyword*) are unchanged since traits provide the methods
  identically.

extension), which verifies body equality before hoisting and resolves the
PSR-4 trait path from composer.json. Verified: php -l clean on all 6 files;
runtime smoke test composing both traits confirms applyKeywordSearch/Exclusion
behave identically to the originals.
