+++ inc/Abilities/Fetch/FetchWordPressApiAbility.php
@@ -12,12 +12,16 @@
 namespace DataMachine\Abilities\Fetch;
 
 use DataMachine\Abilities\PermissionHelper;
+use DataMachine\Abilities\Fetch\Traits\HasApplyKeywordSearch;
+use DataMachine\Abilities\Fetch\Traits\HasApplyKeywordExclusion;
 
 defined( 'ABSPATH' ) || exit;
 
 class FetchWordPressApiAbility {
 
 	use FetchHttpGetTrait;
+	use HasApplyKeywordSearch;
+	use HasApplyKeywordExclusion;
 
 	private static bool $registered = false;
 
@@ -599,51 +603,4 @@ class FetchWordPressApiAbility {
 		return $item_timestamp >= $cutoff;
 	}
 
-	/**
-	 * Apply keyword search filter.
-	 */
-	private function applyKeywordSearch( string $text, string $search_term ): bool {
-		if ( empty( $search_term ) ) {
-			return true;
-		}
-
-		$terms      = array_map( 'trim', explode( ',', $search_term ) );
-		$text_lower = strtolower( $text );
-
-		foreach ( $terms as $term ) {
-			if ( empty( $term ) ) {
-				continue;
-			}
-			if ( strpos( $text_lower, strtolower( $term ) ) !== false ) {
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	/**
-	 * Apply keyword exclusion filter.
-	 *
-	 * Returns true when any of the comma-separated terms in $exclude_keywords
-	 * is present in $text. Inverse of applyKeywordSearch — callers should skip
-	 * items for which this returns true. An empty exclusion list never matches.
-	 */
-	private function applyKeywordExclusion( string $text, string $exclude_keywords ): bool {
-		$exclude_keywords = trim( $exclude_keywords );
-		if ( '' === $exclude_keywords ) {
-			return false;
-		}
-
-		$terms      = array_filter( array_map( 'trim', explode( ',', $exclude_keywords ) ) );
-		$text_lower = strtolower( $text );
-
-		foreach ( $terms as $term ) {
-			if ( strpos( $text_lower, strtolower( $term ) ) !== false ) {
-				return true;
-			}
-		}
-
-		return false;
-	}
 }
